Si succeed

+++ vasper/vasprun.zsh
@@ -8,9 +8,9 @@ function getoptnum()
 {
   # get the opt directory number to make
   local NUM=`find . -mindepth 1 -maxdepth 1 -type d -name opt\* | wc -l`
-  if [ $NUM = 0 ]; then
-    NUM=$(($NUM+1))
-  fi
+  # if [ $NUM = 0 ]; then
+  #   NUM=$(($NUM+1))
+  # fi
   echo $NUM
 }
 
@@ -28,9 +28,11 @@ function makeoptdir()
 {
   # make opt[num] directory
   # if opt* directories exist, remove the last one and remake
-  local OPTNUM=`getoptnum`
-  local DIRNAME="opt${OPTNUM}"
-  rm -rf $DIRNAME
+  ##### $1 : repeat number
+  local DIRNAME="opt${1}"
+  if [ -e "$DIRNAME" ]; then
+    rm -rf $DIRNAME
+  fi
   mkdir $DIRNAME
   cp INCAR $DIRNAME
   cp KPOINTS $DIRNAME
@@ -38,22 +40,23 @@ function makeoptdir()
   if [ "$1" = 1 ]; then
     cp POSCAR $DIRNAME
   else
-    cp "opt$((${OPTNUM}-1))"/CONTCAR $DIRNAME
+    cp "opt$((${1}-1))"/CONTCAR $DIRNAME/POSCAR
   fi
 }
 
 function repeat_setting()
 {
-  ##### $1 : repeat num
-  local COLUMNS=(`head -n 1 vasper_relax.dat | tr -s " "`)
-  local COLUMNS_NUM=$#COLUMNS
-  for i in {1..$COLUMNS_NUM}
+  ##### $1 : repeat number
+  # local COLUMNS=(`head -n 1 vasper_relax.dat | tr -s " "`)
+  COLUMN=(`head -n 1 vasper_relax.dat | tr -s " "`)
+  local COLUMN_NUM=$#COLUMN
+  for i in {2..$COLUMN_NUM}
   do
-    local VAR=`cat vasper_relax.dat | tr -s " " | cut -f ${i} -d " " | sed -n $((${1}+1))`
-    if [ "$COLUMNS[$i]" = "REPEAT" ]; then
+    local VAR=`cat vasper_relax.dat | tr -s " " | cut -f ${i} -d " " | sed -n $((${1}+1))p`
+    if [ "$COLUMN[$i]" = "REPEAT" ]; then
       MAX_REPEAT="$VAR"
     else
-      vasper-makefile.zsh --revise_setting "INCAR" "$COLUMNS[$i]" "$VAR"
+      vasper-makefile.zsh --revise_setting "INCAR" "$COLUMN[$i]" "$VAR"
     fi
   done
 }
@@ -66,7 +69,7 @@ function run()
   function copyfiles()
   {
     ##### $1 : repeat number
-    local FILES=(INCAR CONTCAR OUTCAR vasprun.xml OSZICAR vasper-log.yaml)
+    local FILES=(CONTCAR OUTCAR vasprun.xml OSZICAR vasper-log.yaml)
     for FILE in $FILES
     do
       if [ -e $FILE ]; then
@@ -75,13 +78,6 @@ function run()
     done
   }
 
-  function get_execdir()
-  {
-
-  }
-
-  OPTNUM=`getoptnum`
-  cd opt${OPTNUM}
   cp POSCAR POSCAR_0
   for i in {1..${1}}
   do
@@ -90,12 +86,10 @@ function run()
     if [ ! -e "vasper-log.yaml" ]; then
       if [ -e "CONTCAR" ]; then
         copyfiles $i
-        cd -
         return 0
       else
         echo "there is not COUTCAR and vaspper-log.yaml file"
         echo "some error occured"
-        cd -
         exit 1
       fi
 
@@ -103,12 +97,10 @@ function run()
       copyfiles $i
       local CONVERGE=`vasper-log.py --all | grep converged | rev | cut -f 1 -d " " | rev`
       if [ "$CONVERGE" = "True" ]; then
-        cd -
         return 0
       fi
     fi
   done
   echo "too many repeats"
-  cd -
   exit 1
 }
